+++ surface_target_contract_test.go
@@ -1,0 +1,6 @@
+package wgpu
+
+var (
+	_ func(*Instance, SurfaceTarget) (*Surface, error)       = (*Instance).CreateSurfaceFromTarget
+	_ func(*Instance, SurfaceTargetUnsafe) (*Surface, error) = (*Instance).CreateSurfaceUnsafe
+)